Refactor learner profile into openedx/features directory
LEARNER-1855
