+++ lms/envs/common.py
@@ -1782,7 +1782,6 @@ REQUIRE_JS_PATH_OVERRIDES = {
     'js/courseware/link_clicked_events': 'js/courseware/link_clicked_events.js',
     'js/courseware/toggle_element_visibility': 'js/courseware/toggle_element_visibility.js',
     'js/student_account/logistration_factory': 'js/student_account/logistration_factory.js',
-    'js/student_profile/views/learner_profile_factory': 'js/student_profile/views/learner_profile_factory.js',
     'js/courseware/courseware_factory': 'js/courseware/courseware_factory.js',
     'js/groups/views/cohorts_dashboard_factory': 'js/groups/views/cohorts_dashboard_factory.js',
     'js/groups/discussions_management/discussions_dashboard_factory': 'js/discussions_management/views/discussions_dashboard_factory.js',
@@ -2252,7 +2251,9 @@ INSTALLED_APPS = (
     'openedx.features.course_experience',
     'openedx.features.course_search',
     'openedx.features.enterprise_support',
+    'openedx.features.learner_profile',
 
+    # Experiments
     'experiments',
 
     # DRF filters
