--- lms/static/js/student_profile/models/badges_model.js
@@ -1,7 +1,0 @@
-(function(define) {
-    'use strict';
-    define(['backbone'], function(Backbone) {
-        var BadgesModel = Backbone.Model.extend({});
-        return BadgesModel;
-    });
-}).call(this, define || RequireJS.define);
