+++ lms/static/learner_profile
@@ -1,0 +1,1 @@
+../../openedx/features/learner_profile/static/learner_profile
