+++ lms/static/lms/js/build.js
@@ -37,8 +37,8 @@
             'js/student_account/logistration_factory',
             'js/student_account/views/account_settings_factory',
             'js/student_account/views/finish_auth_factory',
-            'js/student_profile/views/learner_profile_factory',
             'js/views/message_banner',
+            'learner_profile/js/learner_profile_factory',
             'lms/js/preview/preview_factory',
             'support/js/certificates_factory',
             'support/js/enrollment_factory',
