+++ lms/static/lms/js/spec/main.js
@@ -88,9 +88,6 @@
             'js/views/file_uploader': 'js/views/file_uploader',
             'js/views/notification': 'js/views/notification',
             'js/student_account/account': 'js/student_account/account',
-            'js/student_profile/views/learner_profile_fields': 'js/student_profile/views/learner_profile_fields',
-            'js/student_profile/views/learner_profile_factory': 'js/student_profile/views/learner_profile_factory',
-            'js/student_profile/views/learner_profile_view': 'js/student_profile/views/learner_profile_view',
             'js/ccx/schedule': 'js/ccx/schedule',
             'js/views/message_banner': 'js/views/message_banner',
 
@@ -784,14 +781,6 @@
         'js/spec/student_account/password_reset_spec.js',
         'js/spec/student_account/register_spec.js',
         'js/spec/student_account/shoppingcart_spec.js',
-        'js/spec/student_profile/badge_list_container_spec.js',
-        'js/spec/student_profile/badge_list_view_spec.js',
-        'js/spec/student_profile/badge_view_spec.js',
-        'js/spec/student_profile/learner_profile_factory_spec.js',
-        'js/spec/student_profile/learner_profile_fields_spec.js',
-        'js/spec/student_profile/learner_profile_view_spec.js',
-        'js/spec/student_profile/section_two_tab_spec.js',
-        'js/spec/student_profile/share_modal_view_spec.js',
         'js/spec/verify_student/image_input_spec.js',
         'js/spec/verify_student/make_payment_step_view_ab_testing_spec.js',
         'js/spec/verify_student/make_payment_step_view_spec.js',
@@ -803,6 +792,14 @@
         'js/spec/views/file_uploader_spec.js',
         'js/spec/views/message_banner_spec.js',
         'js/spec/views/notification_spec.js',
+        'learner_profile/js/spec/learner_profile_factory_spec.js',
+        'learner_profile/js/spec/views/badge_list_container_spec.js',
+        'learner_profile/js/spec/views/badge_list_view_spec.js',
+        'learner_profile/js/spec/views/badge_view_spec.js',
+        'learner_profile/js/spec/views/learner_profile_fields_spec.js',
+        'learner_profile/js/spec/views/learner_profile_view_spec.js',
+        'learner_profile/js/spec/views/section_two_tab_spec.js',
+        'learner_profile/js/spec/views/share_modal_view_spec.js',
         'support/js/spec/collections/enrollment_spec.js',
         'support/js/spec/models/enrollment_spec.js',
         'support/js/spec/views/certificates_spec.js',
