+++ openedx/features/learner_profile/static/learner_profile/js/models/badges_model.js
@@ -1,0 +1,8 @@
+(function(define) {
+    'use strict';
+
+    define(['backbone'], function(Backbone) {
+        var BadgesModel = Backbone.Model.extend({});
+        return BadgesModel;
+    });
+}).call(this, define || RequireJS.define);
