+++ openedx/features/learner_profile/static/learner_profile/templates/badge_list.underscore
